[ANDROSDK-921-2] Fix test & androidTest
