+++ core/src/main/java/org/hisp/dhis/android/core/common/internal/CommonPackageDIModule.java
@@ -56,4 +56,10 @@ public class CommonPackageDIModule {
                                     DHISVersionManager versionManager) {
         return GenericCallData.create(databaseAdapter, retrofit, resourceHandler, versionManager);
     }
+
+    @Provides
+    @Reusable
+    public DataStatePropagator dataStatePropagator(DataStatePropagatorImpl impl) {
+        return impl;
+    }
 }
